% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_pubyear.R
\name{approximate_pubyear}
\alias{approximate_pubyear}
\title{approximate_pubyear}
\usage{
approximate_pubyear(df)
}
\arguments{
\item{df}{data.frame with publication.year field}
}
\value{
data.frame with publication years estimated
}
\description{
Approximate publication year based on available information
}
\examples{
\dontrun{df2 <- approximate_pubyear(df)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

