% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author_info_table.R
\name{author_info_table}
\alias{author_info_table}
\title{ambiguous_authors_table}
\usage{
author_info_table(...)
}
\arguments{
\item{...}{Arguments to be pased}
}
\value{
Author synonyme data frame with fields 'name' and 'synonyme'
}
\description{
Read table of ambiguous author name synonymes
}
\examples{
aa <- ambiguous_authors_table()
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

