% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_first_index.R
\name{choose_first_index}
\alias{choose_first_index}
\title{Choose first index}
\usage{
choose_first_index(text2, split_w)
}
\arguments{
\item{text2}{string text, vector split_w}

\item{split_w}{split_w}
}
\value{
the string which is first in the text field (from strings in the vector)
}
\description{
Find the first index of the splitting token arrays in a string.
}
\details{
finds the first index of the splitting token arrays in a string
}
\examples{
# choose_first_index("printed and sold by R. Marchbank") #  split_w=c(" and ", " by ") -> " and "
}
\author{
Mika Koistinen, Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

