% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time.R
\name{combine.times}
\alias{combine.times}
\title{combine.times}
\usage{
combine.times(timespan)
}
\arguments{
\item{timespan}{time span}
}
\value{
Combined time span
}
\description{
Combine time information
}
\examples{
\dontrun{combine.times(timespan)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

