% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estc-package.R
\docType{package}
\name{estc-package}
\alias{estc}
\alias{estc-package}
\title{British Library Data Toolkit}
\description{
Brief summary of the estc package
}
\details{
\tabular{ll}{
Package: \tab estc\cr
Type: \tab Package\cr
Version: \tab See sessionInfo() or DESCRIPTION file\cr
Date: \tab 2014\cr
License: \tab BSD-2-clause\cr
LazyLoad: \tab yes\cr
}

R Tools for British Library Data
}
\examples{
library(estc)
}
\author{
Leo Lahti, Niko Ilomaki, Mikko Tolonen \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
\url{http://ropengov.github.com/estc}
}
\keyword{package}

