% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate_document_items.R
\name{estimate_document_items}
\alias{estimate_document_items}
\title{estimate_document_items}
\usage{
estimate_document_items(df)
}
\arguments{
\item{df}{data frame}
}
\value{
Vector of estimated number of items for each document.
}
\description{
Estimate number of physical items for a document
}
\examples{
# x <- estimate_document_items(df)
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

