% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pagecount.R
\name{estimate_document_parts}
\alias{estimate_document_parts}
\title{estimate_document_parts}
\usage{
estimate_document_parts(df)
}
\arguments{
\item{df}{data frame}
}
\value{
Vector of estimated number of parts for each document.
}
\description{
Estimate document parts
}
\examples{
# x <- estimate_document_parts(df)
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

