% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geocoordinates.R
\name{get_geocoordinates}
\alias{get_geocoordinates}
\title{get_geocoordinates}
\usage{
get_geocoordinates(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
data.frame
}
\description{
Geocoordinate table
}
\examples{
# x2 <- get_geocoorcinates()
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

