% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geonames.R
\name{get_geonames}
\alias{get_geonames}
\title{get_geonames}
\usage{
get_geonames(dataset = "cities1000", download.dir = NULL)
}
\arguments{
\item{dataset}{Name of the geonames data set to download (for instance "cities1000")}

\item{download.dir}{Optional. Specify download directory.}
}
\value{
Polished data frame
}
\description{
Download geonames data sets.
}
\details{
For the data fields descriptions, see: http://download.geonames.org/export/dump/
}
\examples{
\dontrun{geonames <- get_geonames("cities1000")}
}
\references{
See citation("estc")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
