% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialcases.R
\name{handle_special_cases}
\alias{handle_special_cases}
\title{handle_special_cases}
\usage{
handle_special_cases(x)
}
\arguments{
\item{x}{vector}
}
\value{
polished vector
}
\description{
Handle special cases in preprocessing
}
\examples{
\dontrun{x2 <- handle_special_cases(x)}
}
\references{
See citation("estc")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
