% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_pages_specialcases.R
\name{harmonize_pages_specialcases}
\alias{harmonize_pages_specialcases}
\title{Harmonize pages specialcases}
\usage{
harmonize_pages_specialcases(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
Polished vector
}
\description{
Harmonize pages (special cases)
}
\examples{
# harmonize_pages_specialcases("Caption title 2")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

