% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{harmonize_time_info}
\alias{harmonize_time_info}
\title{harmonize_time_info}
\usage{
harmonize_time_info(x)
}
\arguments{
\item{x}{Vector (time field)}
}
\value{
Polished field
}
\description{
Harmonize time information
}
\details{
Harmonize month names
}
\examples{
\dontrun{x2 <- harmonize_time_info(x)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

