% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_geonames.R
\name{match_geonames}
\alias{match_geonames}
\title{Match place names to geonames database}
\usage{
match_geonames(places, geonames)
}
\arguments{
\item{places}{Vector of place names to match}

\item{geonames}{Geonames table}
}
\value{
Geonames names
}
\description{
Match place names to geonames database
}
\examples{
# TODO
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

