% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.fields.R
\name{order.fields}
\alias{order.fields}
\title{order.fields}
\usage{
order.fields(df, start.field, end.field)
}
\arguments{
\item{df}{data.frame}

\item{start.field}{First field name}

\item{end.field}{Last field name}
}
\value{
data.frame with ordered fields
}
\description{
Put fields in correct order
}
\examples{
\dontrun{df2 <- order.fields(df, start.field, end.field)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

