% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_print_fields.R
\name{pick_print_fields}
\alias{pick_print_fields}
\title{Pick print fields}
\usage{
pick_print_fields(x, fields2)
}
\arguments{
\item{x}{list of strings}

\item{fields2}{c(string_field1, string_field2...)}
}
\value{
for each x, the string after last field that has been found, else the string after first one, else NA
}
\description{
Pick print fields from publisher vector.
}
\examples{
\dontrun{x2 <- pick_print_fields(x, field)}
}
\references{
See citation("estc")
}
\author{
Mika Koistinen, Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
