% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_shortest_text.R
\name{pick_shortest_text}
\alias{pick_shortest_text}
\title{Pick shorter text}
\usage{
pick_shortest_text(vec)
}
\arguments{
\item{vec}{vector of strings}
}
\value{
shortest string
}
\description{
returns text which is shortest from vector of strings
}
\references{
See citation("bibliographica")
}
\author{
Mika Koistinen, Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
