% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time.R
\name{polish.650y}
\alias{polish.650y}
\title{polish.650y(x)}
\usage{
polish.650y(x)
}
\arguments{
\item{x}{a vector of field 650y}
}
\value{
The polished vector
}
\description{
Polish the field 650y
}
\examples{
\dontrun{x2 <- polish.650y(x)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

