% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/author.R
\name{polish_author_duplicates}
\alias{polish_author_duplicates}
\title{polish_author_duplicates}
\usage{
polish_author_duplicates(authors, author.synonyme.table = NULL)
}
\arguments{
\item{authors}{Vector of author names}

\item{author.synonyme.table}{Author synonyme data frame with fields 'name' and 'synonyme'}
}
\value{
Vector of harmonized author names
}
\description{
Harmonize synonymous author ids
}
\examples{
\dontrun{s2 <- polish_author_duplicates(s)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

