% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publisher.R
\name{polish_publisher2}
\alias{polish_publisher2}
\title{polish_publisher2}
\usage{
polish_publisher2(x)
}
\arguments{
\item{x}{vector}
}
\value{
Polished vector
}
\description{
Polish publisher field
}
\details{
Polish publisher field. As a side operation, write the polished summary data in a file
}
\examples{
# polish_publisher("printed and sold by R. Marchbank")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

