% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ecco.R
\name{read_ecco}
\alias{read_ecco}
\title{Read ECCO}
\usage{
read_ecco(version = 2)
}
\arguments{
\item{version}{Version number 1: ECCOI; 2: ECCOII}
}
\value{
ECCO data.frame
}
\description{
Read ECCO data dump.
}
\details{
Assumes that the working directory includes the source data file.
}
\examples{
\dontrun{ecco <- read_ecco(version = 2)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

