% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/est.R
\name{read_ESTC_data}
\alias{read_ESTC_data}
\title{read_ESTC_data}
\usage{
read_ESTC_data(file)
}
\arguments{
\item{file}{Source ESTC data file}
}
\value{
data.frame
}
\description{
Read parsed British Library ESTC data
}
\examples{
\dontrun{est <- read_ESTC_data(file)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

