% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ending_chars.R
\name{remove_ending_chars}
\alias{remove_ending_chars}
\title{Remove Ending Chars}
\usage{
remove_ending_chars(x)
}
\arguments{
\item{x}{vector of strings}
}
\value{
vector cleaned
}
\description{
remove ending characters (and , ;) from every string in a vector.
}
\details{
removes ending characters and , ; from every string in a vector.
}
\examples{
# remove_ending_chars("printed and sold by R. Marchbank,")
}
\references{
See citation("bibliographica")
}
\author{
Mika Koistinen \email{leo.lahti@iki.fi}
}
\keyword{utilities}
