% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_doer_and_place.R
\name{split_doer_and_place}
\alias{split_doer_and_place}
\title{Split doer and place}
\usage{
split_doer_and_place(data)
}
\arguments{
\item{data}{vector of strings}
}
\value{
list(x_doer,x_place)
}
\description{
Splits string by the first index values of split_w <- c(" in "," at ",", at ","; at ", ",at", ",in", ", in",", near ",",no. ",", no. ").
}
\details{
split_doer_and_place
}
\examples{
# split_does_and_place("printed and sold by R. Marchbank")
}
\author{
Mika Koistinen, Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

