% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonyme_list2df.R
\name{synonyme_list2df}
\alias{synonyme_list2df}
\title{Synonyme list to data.frame}
\usage{
synonyme_list2df(synonyme.list, sep = ",", include.lowercase = FALSE)
}
\arguments{
\item{synonyme.list}{synonyme list}

\item{sep}{separator}

\item{include.lowercase}{include lowercase}
}
\value{
data.frame
}
\description{
Synonyme list to data.frame.
}
\examples{
# TODO
}
\references{
See citation("estc")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
