% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edition.R
\name{clean.edition}
\alias{clean.edition}
\title{clean.edition}
\usage{
clean.edition(x)
}
\arguments{
\item{x}{Edition field (a vector)}
}
\value{
Polished edition info
}
\description{
Clean edition
}
\examples{
clean.edition("seventh edition")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

