% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/topic.R
\name{polish.topic}
\alias{polish.topic}
\title{polish.topic}
\usage{
polish.topic(x)
}
\arguments{
\item{x}{A vector with topics}
}
\value{
A vector with topics polished
}
\description{
Polish topic
}
\examples{
\dontrun{x2 <- polish.topic(x)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

