% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geo.R
\name{polish_geography}
\alias{polish_geography}
\title{polish_geography}
\usage{
polish_geography(x)
}
\arguments{
\item{x}{A vector of geographic names}
}
\value{
Polished vector
}
\description{
Polish geography
}
\examples{
\dontrun{x2 <- polish_geography(x)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

